+++ packages/apptive_grid_form/test/geolocation_form_widget_test.dart
@@ -1574,7 +1574,7 @@ void main() {
         throw 'MissingMockResponse for GET $requestUri';
       });
 
-      when(() => apptiveGridClient.options).thenReturn(options);
+      apptiveGridClient.options = options;
       when(() => apptiveGridClient.sendPendingActions())
           .thenAnswer((_) async => []);
       when(